fix(e2e): prevent declaration/compliance test interleaving race condition
Declaration tests failed because Playwright can interleave chromium and
compliance projects (both depend only on setup). When a compliance
beforeAll sets status='submitted', declaration step 1 redirects to step 6.

Extract DB reset logic into shared helper, add beforeAll to declaration
suite, and increase logout test timeout to 120s.
